delete acl

+++ packages/alea-frontend/pages/api/access-control/delete-acl.ts
@@ -3,6 +3,7 @@ import {
   checkIfPostOrSetError,
   executeTxnAndEndSet500OnError,
 } from '../comment-utils';
+import { isCurrentUserMemberOfAClupdater } from '../acl-utils/acl-common-utils';
 
 export default async function handler(
   req: NextApiRequest,
@@ -12,7 +13,8 @@ export default async function handler(
   // TODO: Check if from updaterACLId
   const id = req.body.id as string;
   if (!id || typeof id !== 'string') return res.status(422).send('Missing id.');
-
+  if (!await isCurrentUserMemberOfAClupdater(id, res, req))
+    return res.status(403).end();
   await executeTxnAndEndSet500OnError(
     res,
     'DELETE FROM AccessControlList WHERE id=?',
